+++ integrationtests/fips/fips_test.go
@@ -1,5 +1,3 @@
-//go:build go1.26
-
 package fips
 
 import (
