--- internal/handshake/fips140_go126.go
@@ -1,9 +1,0 @@
-//go:build go1.26
-
-package handshake
-
-import "crypto/fips140"
-
-func withoutFIPSEnforcement(f func()) {
-	fips140.WithoutEnforcement(f)
-}
