--- internal/handshake/fips140_legacy.go
@@ -1,7 +1,0 @@
-//go:build !go1.26
-
-package handshake
-
-func withoutFIPSEnforcement(f func()) {
-	f()
-}
